+++ app/mailers/pay/user_mailer.rb
@@ -1,8 +1,8 @@
 module Pay
   class UserMailer < ApplicationMailer
     def receipt
-      if params[:charge].respond_to? :receipt
-        attachments[params[:charge].filename] = params[:charge].receipt
+      if params[:pay_charge].respond_to? :receipt
+        attachments[params[:pay_charge].filename] = params[:pay_charge].receipt
       end
 
       mail to: to