+++ lib/pay/stripe/webhooks/charge_refunded.rb
@@ -5,8 +5,8 @@ module Pay
         def call(event)
           pay_charge = Pay::Stripe::Charge.sync(event.data.object.id, stripe_account: event.try(:account))
 
-          if pay_charge && Pay.send_emails
-            Pay::UserMailer.with(pay_customer: pay_charge.customer, charge: pay_charge).refund.deliver_later
+          if pay_charge && Pay.send_email?(:refund, pay_charge)
+            Pay::UserMailer.with(pay_customer: pay_charge.customer, pay_charge: pay_charge).refund.deliver_later
           end
         end
       end